+++ test/Unit/AbstractNormalizerTestCase.php
@@ -13,9 +13,9 @@ declare(strict_types=1);
 
 namespace Localheinz\Composer\Json\Normalizer\Test\Unit;
 
+use Ergebnis\Json\Normalizer\Json;
+use Ergebnis\Json\Normalizer\NormalizerInterface;
 use Ergebnis\Test\Util\Helper;
-use Localheinz\Json\Normalizer\Json;
-use Localheinz\Json\Normalizer\NormalizerInterface;
 use PHPUnit\Framework;
 
 /**